Rename "notifications bar" → "notification bar"
