+++ source/refined-github.ts
@@ -171,7 +171,7 @@ import './features/clone-branch';
 import './features/deep-reblame';
 import './features/clear-pr-merge-commit-message';
 import './features/go-to-action-from-file';
-import './features/stop-redirecting-in-notifications-bar';
+import './features/stop-redirecting-in-notification-bar';
 
 // Add global for easier debugging
 (window as any).select = select;
